fix: avoid division by zero in inclusive kinematics DA method (#1005)
### Briefly, what does this PR introduce?
This avoids a division by zero when only the scattered electron is in
the final state, and no other hadrons are present. In that case `Esum`
and `pzsum` are both still exactly zero, and `sigma_h` is still exactly
zero as well.

### What kind of change does this PR introduce?
- [x] Bug fix (issue:
https://github.com/eic/EICrecon/actions/runs/6247530971/job/16961738954?pr=987#step:5:1059)
- [ ] New feature (issue #__)
- [ ] Documentation update
- [ ] Other: __

### Please check if this PR fulfills the following:
- [ ] Tests for the changes have been added
- [ ] Documentation has been added / updated
- [ ] Changes have been communicated to collaborators

### Does this PR introduce breaking changes? What changes might users
need to make to their code?
No.

### Does this PR change default behavior?
No.