feat: add two-stage test job (digitization/reconstruction); write EventHeader by default (#1014)
### Briefly, what does this PR introduce?
This adds a test job that runs digitization and reconstruction in
separate eicrecon steps, with only a raw file artifact as communication
between them. This should allow for ensuring #667 is continuously
tested.

It is likely that this test won't work until some additional work is
completed either here or elsewhere, depending on the scope of that extra
work.

The test is limited to the barrel ecal because we currently have to
specify all required raw collections as command line arguments, and this
seems representative enough. Tracking systems are more collections, and
none of the other calorimeter systems have deeply involved experts, so
that's how I ended up with barrel ecal.

Quality of life ideas:
- easier way to specify output selection from the default list, e.g. by
regex (I would have picked `.*Raw.*` if that was possible)

### What kind of change does this PR introduce?
- [x] Bug fix (issue #667)
- [ ] New feature (issue #__)
- [ ] Documentation update
- [ ] Other: __

### Please check if this PR fulfills the following:
- [x] Tests for the changes have been added
- [ ] Documentation has been added / updated
- [ ] Changes have been communicated to collaborators

### Does this PR introduce breaking changes? What changes might users
need to make to their code?
No.

### Does this PR change default behavior?
No.

---------

Co-authored-by: Dmitry Kalinkin <dmitry.kalinkin@gmail.com>