ci: ccache max size increase and show stats; key name changes (#774)
### Briefly, what does this PR introduce?
For debug builds the previous 500MB doesn't fit the entire build so we
end up rebuilding all the time. This bumps it up to 1500MB and adds a
stats summary at the end which should allow us to assess how many misses
we have.

For more efficient lookups in branches off branches, this explicitly
adds the PR head and base branch to the lookup list, and changes the
ccache key to the head branch for PRs. This means the temporary merge
ref "774/merge" won't be used as the key, since that cannot be
determined by dependent branches.

### What kind of change does this PR introduce?
- [x] Bug fix (issue #__)
- [x] New feature (issue #__)
- [ ] Documentation update
- [ ] Other: __

### Please check if this PR fulfills the following:
- [ ] Tests for the changes have been added
- [ ] Documentation has been added / updated
- [ ] Changes have been communicated to collaborators

### Does this PR introduce breaking changes? What changes might users
need to make to their code?
No.

### Does this PR change default behavior?
No.