fix: address clang-tidy clang-analyzer warnings (#1013)
### Briefly, what does this PR introduce?
This addresses nearly all issues highlighted in #757. Mainly essentially
unused variables that the compiler didn't already flag because of
conditional assignment patterns.

This doesn't fix the warning about the bypassing of virtual dispatch in
the RichGeoDRICH and RichGeoPFRICH constructors. Maybe prefixing the
class names makes it clear we don't want dispatch and shuts up the
warning, but needs testing. (cc @chchatte92 @c-dilks )

### What kind of change does this PR introduce?
- [x] Bug fix (issue #757)
- [ ] New feature (issue #__)
- [ ] Documentation update
- [ ] Other: __

### Please check if this PR fulfills the following:
- [ ] Tests for the changes have been added
  - These are all tests not available in the clang-tidy-14 we're using.
- [ ] Documentation has been added / updated
- [ ] Changes have been communicated to collaborators

### Does this PR introduce breaking changes? What changes might users
need to make to their code?
No.

### Does this PR change default behavior?
No.