Reducing number of track projections inside dRICH (#999)
### Briefly, what does this PR introduce?
Currently we take 35 points inside the dRICH for the track propagation
per charged particle. This increases the CPU time. This PR (a draft)
includes 25 points instead of 35 points, 5 points for aerogel and 20
points for the gas. This draft PR is in response to the suggestion:
https://github.com/eic/EICrecon/pull/991#issuecomment-1726523966

### What kind of change does this PR introduce?
- [ ] Bug fix (issue #__)
- [ ] New feature (issue #__)
- [ ] Documentation update
- [X] Other: __

### Please check if this PR fulfills the following:
- [ ] Tests for the changes have been added
- [ ] Documentation has been added / updated
- [X] Changes have been communicated to collaborators
This Draft PR is in response to the above mentioned comments from Chris
Dilks.

### Does this PR introduce breaking changes? What changes might users
need to make to their code?
No. 
### Does this PR change default behavior?
No.

---------

Co-authored-by: Christopher Dilks <c-dilks@users.noreply.github.com>
Co-authored-by: Dmitry Kalinkin <dmitry.kalinkin@gmail.com>