fix: remove all odr violations, and enforce them in ASAN_OPTIONS (#1023)
### Briefly, what does this PR introduce?
This removes the remaining [one-definition rule
(odr)](https://en.cppreference.com/w/cpp/language/definition) violations
in the CKFTracking (tracking.so plugin and library), and the reco.so
plugin and library. It also renames the misnamed janadot.cc to
janatop.cc. This now allows us to turn on the detect_odr_violations in
ASAN_OPTIONS.

The next steps could be to remove the flags PLUGIN_USE_CC_ONLY and
PLUGIN_GLOB_ALL again, and default to PLUGIN_USE_CC_ONLY (which would
make plugin and library based on entirely distinct source files).
However, that's not quite working yet (for reasons I have to understand
and resolve, are limited to services, and which don't introduce odr
violations). In the mean time we should push this into main and prevent
breakage from getting added inadvertently.

### What kind of change does this PR introduce?
- [x] Bug fix (issue #__)
- [ ] New feature (issue #__)
- [ ] Documentation update
- [ ] Other: __

### Please check if this PR fulfills the following:
- [x] Tests for the changes have been added
- [ ] Documentation has been added / updated
- [ ] Changes have been communicated to collaborators

### Does this PR introduce breaking changes? What changes might users
need to make to their code?
No.

### Does this PR change default behavior?
No.