fix: rm independent geometry reading in RICHGeo (#1036)
### Briefly, what does this PR introduce?
This removes the unused (deprecated) IrtGeo interfaces that load their
own compact file. This is in particular relevant because RichGeo cleans
up the geometry regardless with a `destroyInstance` that is applied on
both the DD4hep_service and RichGeo_service geometries. That could
potentially cause a race condition when both DD4hep_service and
RichGeo_service try to destroy the same geometry instance (it probably
is happening already but is caught by the try catch around both
instances). This change also allows us to move towards more const
correctness in the treatment of DD4hep geometries (i.e. not passing the
geometry as a non-const pointer).

### What kind of change does this PR introduce?
- [x] Bug fix (issue: `destroyInstance` in RichGeo_service also applies
on DD4hep_service geometry)
- [ ] New feature (issue #__)
- [ ] Documentation update
- [ ] Other: __

### Please check if this PR fulfills the following:
- [ ] Tests for the changes have been added
- [ ] Documentation has been added / updated
- [x] Changes have been communicated to collaborators @c-dilks
@chchatte92

### Does this PR introduce breaking changes? What changes might users
need to make to their code?
No.

### Does this PR change default behavior?
No.