Enable user configurability of jet reconstruction (#1025)
### Briefly, what does this PR introduce?

This PR updates the interfaces to jet reconstruction algorithm so as to
allow users to change parameters of the algorithm from the command line,
via a configuration file, or through initializer lists. For example:

`eicrecon -Preco:ReconstructedJets:minCstPt=2`

will raise the minimum pT of objects fed to the jet algorithm to 2 GeV/c
for the reconstructed jet collection.

The utilized jet algorithm, recombination scheme, and area type can also
be set in the same manner. If any of these are set to an unknown value
(e.g. `-Preco:ReconstructedJets:jetAlgo=mystery_algo`), the EICrecon
algorithm will issue a warning and run with a default value instead.

### What kind of change does this PR introduce?
- [ ] Bug fix (issue #__)
- [X ] New feature (issue #901 )
- [ ] Documentation update
- [ ] Other: __

### Please check if this PR fulfills the following:
- [ ] Tests for the changes have been added
- [ ] Documentation has been added / updated
- [X ] Changes have been communicated to collaborators

### Does this PR introduce breaking changes? What changes might users
need to make to their code?

No.

### Does this PR change default behavior?

No.