Fix crashes in JFactoryPodioT::ClearData() (#1008)
We are observing inconsistent crashes in delete operator invocation in
ClearData() on certain events. These faults occur on macOS with a high
rate, specifically for CalorimeterHitContribution. We also have reports
of some percent of reconstruction jobs (on Linux) crashing specifically
on ReconstructedParticle.

There might be an issue with PODIO object clones, they should be
pointing to the same object data. There doesn't seem to be any facility
for reference counting.

A slight complication with addressing this is that the failure comes in
the implementation FactoryPodioT from JANA2 libraries. That factory is
implicitly instantiated as a result of EICrecon calling the
`JEvent::InsertCollectionAlreadyInFrame` method in `JEventSourcePODIO`.
This PR switches to EICrecon variant of the same code to have a chance
of handling the issue locally.

It is not clear if mData is used or just populated for compatibility.
This is an attempt at disabling the functionality in order to mitigate
the issue.

Resolves: #949
cc @wdconinc @rahmans1