CalorimeterHitDigi: add photon counter based readout
This implements modes to add additional signal smearing associated with
the scintillation process in calorimeters. The SiPM mode allows to study
effect of saturation for a given material, light collection efficiency
and type of sensor.

The part with binomial distribution sampling is a SiPM formalism from
Jin Huang's work done for sPhenix:
https://github.com/sPHENIX-Collaboration/coresoftware/commit/f746f130c11adfdb70b1978a5d3ebf1b258fe1e8