feat: add system and layer ID to Acts::Surface GeometryIdentifier (#1057)
### Briefly, what does this PR introduce?
The Acts::Surface geometry identifier is intended to store information
we are currently passing through two other lists in the TrackPropagation
factory. This fills the geometry identifier field with:
- the 'surface' as layer ID,
- the 'detector' as extra ID (which is inteded for subsystem
identifiers).

These changes will allow us to get rid of the extra vectors that are
passed to the algorithm, which in turn allows us to use the same
algorithm for real and constructed surfaces.

### What kind of change does this PR introduce?
- [ ] Bug fix (issue #__)
- [x] New feature (issue #__)
- [ ] Documentation update
- [ ] Other: __

### Please check if this PR fulfills the following:
- [ ] Tests for the changes have been added
- [ ] Documentation has been added / updated
- [ ] Changes have been communicated to collaborators

### Does this PR introduce breaking changes? What changes might users
need to make to their code?
No.

### Does this PR change default behavior?
No.