feat: enable ubsan by default in CI, with print_stacktrace=1 and suppressions (#987)
This enables ubsan by default in CI, with some more undefined behaviors
(float-divide-by-zero, unsigned-integer-overflow, local-bounds). It has
two suppressions already, but more will no doubt be needed to limit the
output to what we control. There is no way to have an exitcode != 0 when
undefined behavior is found, so this depends a bit on people checking
the output periodically.

TODO:
- [x] #989

---------

Co-authored-by: Dmitry Kalinkin <dmitry.kalinkin@gmail.com>