fix: support fmt v10 and C++20 (#1073)
### Briefly, what does this PR introduce?
These are the first commits in #1072 which should be entirely compatible
with the current version of Acts v21.1.1, fmt v9, and C++17.

Note: this adds C++20 compatibility in EICrecon code (C++20 adds
constexpr support for several fmt functions, which imposes conditions on
the inputs we pass to them), but it does not mean that this PR allows
one to compile EIC with C++20. The Acts dependency of v21.1.1 only
allows C++17 support.

### What kind of change does this PR introduce?
- [x] Bug fix (issue: not compatible with EICrecon)
- [ ] New feature (issue #__)
- [ ] Documentation update
- [ ] Other: __

### Please check if this PR fulfills the following:
- [ ] Tests for the changes have been added
- [ ] Documentation has been added / updated
- [ ] Changes have been communicated to collaborators

### Does this PR introduce breaking changes? What changes might users
need to make to their code?
No.

### Does this PR change default behavior?
No.