feat: depend on algorithms; demo interface in CaloClusterRecoCoG (#1085)
### Briefly, what does this PR introduce?
This adds support for [algorithms](https://github.com/eic/algorithms) in
EICrecon. It depends on getting the updated algorithms@main into
eic-shell, to get the addition of `find_package(algorithms)` support.
This does not add any 'real' algorithms support but demonstrates the
accessibility of the interface in CalorimeterClusterRecoCoG.

### What kind of change does this PR introduce?
- [ ] Bug fix (issue #__)
- [x] New feature (issue #__)
- [ ] Documentation update
- [ ] Other: __

### Please check if this PR fulfills the following:
- [ ] Tests for the changes have been added
- [ ] Documentation has been added / updated
- [x] Changes have been communicated to collaborators @sly2j 

### Does this PR introduce breaking changes? What changes might users
need to make to their code?
No.

### Does this PR change default behavior?
No.