fix(ci): upload build/ as tar.gz for speedup (#1112)
### Briefly, what does this PR introduce?
The way GitHub actions upload-artifact uploads each file individually,
and stores them in a compressed archive on the server. This means that
there are 1500+ files uploaded, each counting as an API call. It's also
slow due to the way the protocols work. Compressing on the node, then
uploading one file could be net faster.

### What kind of change does this PR introduce?
- [x] Bug fix (issue: we spend more time uploading than compiling)
- [ ] New feature (issue #__)
- [ ] Documentation update
- [ ] Other: __

### Please check if this PR fulfills the following:
- [ ] Tests for the changes have been added
- [ ] Documentation has been added / updated
- [ ] Changes have been communicated to collaborators

### Does this PR introduce breaking changes? What changes might users
need to make to their code?
No.

### Does this PR change default behavior?
No.