feat: upgrade EICrecon to Acts v30 (#1072)
### Briefly, what does this PR introduce?
This upgrades EICrecon to work with Acts v30.3.2.

Because of the way changes to Acts are implemented, it is often beyond
cumbersome to achieve backwards compatibility with preprocessor
directives. While in a model of frequent updates this may be more
achievable in the interest of regression testing, it is not a good use
of anyone's time to implement this for large version jumps.

TODO:
- [ ] revert back to default jug_xl:nightly

### What kind of change does this PR introduce?
- [ ] Bug fix (issue #__)
- [x] New feature (issue #623)
- [ ] Documentation update
- [ ] Other: __

### Please check if this PR fulfills the following:
- [ ] Tests for the changes have been added
- [ ] Documentation has been added / updated
- [ ] Changes have been communicated to collaborators

### Does this PR introduce breaking changes? What changes might users
need to make to their code?
Yes. This will require Acts v30.0.0, compiled with examples
(`ACTS_BUILD_EXAMPLES=ON`).

### Does this PR change default behavior?
No changes in EICrecon which would result in different output were
intentionally introduced. If there are logic changes (e.g. CKFTracking
now fitting per seed), they were following changes in the upstream
interfaces.