services/io/podio: rely on ::collection_type feature from recent PODIO (#1108)
This takes advantage of https://github.com/AIDASoft/podio/pull/465 to
start deprecating one of the two responsibilities of the
`make_datamodel_glue.py`. A similar change to JANA2 will be done
independently. I suggest to keep the `#if` pragmas for now, in case we
want to bisect possible issues with the latest PODIO.