fix: update CMAKE_CXX_STANDARD default to 20, now required by Acts (#1129)
### Briefly, what does this PR introduce?
This updates the default C++ standard from 17 to 20, which is now
required by Acts. As of now, there do not appear to be any Acts headers
that are using C++20 concepts (pun intended) but that should be
considered a coincidence. After this PR, C++20 constructs in EICrecon
are officially allowed.

I thought about restricting it to just C++20, not allowing anything
higher, but that prevents us from implementing C++23 support through the
regular development process.

### What kind of change does this PR introduce?
- [x] Bug fix (issue: we're running with C++17 standard, not enforced,
and Acts headers may have C++20 features)
- [ ] New feature (issue #__)
- [ ] Documentation update
- [ ] Other: __

### Please check if this PR fulfills the following:
- [ ] Tests for the changes have been added
- [ ] Documentation has been added / updated
- [ ] Changes have been communicated to collaborators

### Does this PR introduce breaking changes? What changes might users
need to make to their code?
No.

### Does this PR change default behavior?
No.