fix: link tests with --no-as-needed due to JANA #255 (#1130)
### Briefly, what does this PR introduce?
This fixes an issue in the linking of algorithms_test which popped up
today (and likely was always there...). Until
https://github.com/JeffersonLab/JANA2/pull/255 is included, we install a
libJANA.so with undefined references to podio because it doesn't include
it in the NEEDED section. By passing `--no-as-needed` we don't strip the
podio calls from the algorithms_test executables where they are unused.

### What kind of change does this PR introduce?
- [x] Bug fix (issue #__)
- [ ] New feature (issue #__)
- [ ] Documentation update
- [ ] Other: __

### Please check if this PR fulfills the following:
- [ ] Tests for the changes have been added
- [ ] Documentation has been added / updated
- [ ] Changes have been communicated to collaborators

### Does this PR introduce breaking changes? What changes might users
need to make to their code?
No.

### Does this PR change default behavior?
No.

---------

Co-authored-by: Dmitry Kalinkin <dmitry.kalinkin@gmail.com>