fix: pass TrackerHitCollection to TrackerMeasurement, not std::vector<T*> (#1146)
### Briefly, what does this PR introduce?
Instead of passing a std::vector<T*> to the TrackerMeasurementFromHits
algorithm, we should pass a TrackerHitCollection. This PR modifies the
algorithm and factory. The factory accepts an array of input collections
in a similar way as the TrackerHitCollector does, but that functionality
is not currently used (since we use the TrackerHitCollector).
Nevertheless, it means we also need to use a subset collection here.

### What kind of change does this PR introduce?
- [x] Bug fix (issue #1145)
- [ ] New feature (issue #__)
- [ ] Documentation update
- [ ] Other: __

### Please check if this PR fulfills the following:
- [ ] Tests for the changes have been added
- [ ] Documentation has been added / updated
- [ ] Changes have been communicated to collaborators

### Does this PR introduce breaking changes? What changes might users
need to make to their code?
No.

### Does this PR change default behavior?
No.