fix(ci): install into prefix and upload as tar artifact (#1125)
### Briefly, what does this PR introduce?
Similar to #1112, this PR installs EICrecon into a separate
CMAKE_INSTALL_PREFIX (to ensure the rest of the pipeline only relies on
explicitly installed targets, not things that are part of the repository
but not installed). It uploads a single (uncompressed, see #1124) tar
archive with the install directory, ensuring that file permissions are
maintained and that we don't hit our API limits.

### What kind of change does this PR introduce?
- [x] Bug fix (issue #__)
- [ ] New feature (issue #__)
- [ ] Documentation update
- [ ] Other: __

### Please check if this PR fulfills the following:
- [ ] Tests for the changes have been added
- [ ] Documentation has been added / updated
- [ ] Changes have been communicated to collaborators

### Does this PR introduce breaking changes? What changes might users
need to make to their code?
No.

### Does this PR change default behavior?
No.

---------

Co-authored-by: pre-commit-ci[bot] <66853113+pre-commit-ci[bot]@users.noreply.github.com>