fix(ci): run clang-tidy with extra-args -std=c++20 (#1150)
### Briefly, what does this PR introduce?
This should address the clang-tidy issues on main (and likely soon in a
PR). clang-tidy runs without errors in my local eic-shell test setup
now.

I considered adding the logic to allow the builds to be against a
variable that encodes the C++ standard, but that quickly ran out of hand
(see discussion on doing this for compilers). With one standard every
three years, I think it's reasonable to deal with this as we go. We only
support C++20 now anyway.

### What kind of change does this PR introduce?
- [x] Bug fix (issue #1149)
- [ ] New feature (issue #__)
- [ ] Documentation update
- [ ] Other: __

### Please check if this PR fulfills the following:
- [ ] Tests for the changes have been added
- [ ] Documentation has been added / updated
- [ ] Changes have been communicated to collaborators

### Does this PR introduce breaking changes? What changes might users
need to make to their code?
No.

### Does this PR change default behavior?
No.