fix: set C++std to 20, allow higher; agree with Acts and ROOT (#1135)
### Briefly, what does this PR introduce?
Maybe this is a better approach too: set by default to 17, allow user to
set higher, check that C++std agrees with what ROOT and Acts are
assuming.

Question: should we automatically upgrade default 17 to 20 if we find
that Acts or ROOT requires it?

### What kind of change does this PR introduce?
- [x] Bug fix (issue: how to support C++17 and 20 at the same time)
- [ ] New feature (issue #__)
- [ ] Documentation update
- [ ] Other: __

### Please check if this PR fulfills the following:
- [ ] Tests for the changes have been added
- [ ] Documentation has been added / updated
- [ ] Changes have been communicated to collaborators

### Does this PR introduce breaking changes? What changes might users
need to make to their code?
No.

### Does this PR change default behavior?
No.

---------

Co-authored-by: Dmitry Kalinkin <dmitry.kalinkin@gmail.com>