feat: use algorithms:: interface CalorimeterHitDigi (incl RandomSvc) (#1148)
### Briefly, what does this PR introduce?
In the ongoing attempts at making the algorithms self-descriptive and
conformant to the algorithms:: specification, this converts another one:
CalorimeterHitDigi. This wasn't as straightforward as
CalorimeterClusterRecoCoG because of the RandomSvc. No attempt is made
to connect a JANA2 random service (in part because I don't think JANA2
has one; @nathanwbrei is that still true?). Due to the change of random
generator, this is not expected to agree bit-for-bit with what is in
main.

### What kind of change does this PR introduce?
- [ ] Bug fix (issue #__)
- [x] New feature (issue #__)
- [ ] Documentation update
- [ ] Other: __

### Please check if this PR fulfills the following:
- [x] Tests for the changes have been added:
algorithms_test/calorimetry_CalorimeterHitDigi.cc needed some
modifications to deal with the unique pointer memory model.
- [ ] Documentation has been added / updated
- [x] Changes have been communicated to collaborators @sly2j 

### Does this PR introduce breaking changes? What changes might users
need to make to their code?
No.

### Does this PR change default behavior?
No.