fix: install {asan,lsan,tsan,ubsan}.supp with CMake; other CMake cleanup (#1140)
### Briefly, what does this PR introduce?
This installs the {asan,lsan,tsan,ubsan}.supp files into
`share/EICrecon` when the installation has support for the respective
sanitizers. At this point I'm leaving the *.supp files in .github/ since
we don't have an equivalent other directory, and with the current
practice of in-tree installations we don't really want to use share/
itself.

It also applies a few minor CMake cleanups:
- use `message(VERBOSE)` instead of an EICRECON flag; one can use `cmake
--log-level VERBOSE` to print this,
- include GNUInstallDirs for consistent directory support (not yet
propagated to plugins)
- use install DIRECTORY with FILE_PATTERN for cmake/*.cmake files
instead of a GLOB

### What kind of change does this PR introduce?
- [x] Bug fix (issue #1126)
- [ ] New feature (issue #__)
- [ ] Documentation update
- [ ] Other: __

### Please check if this PR fulfills the following:
- [ ] Tests for the changes have been added
- [ ] Documentation has been added / updated
- [ ] Changes have been communicated to collaborators

### Does this PR introduce breaking changes? What changes might users
need to make to their code?
No.

### Does this PR change default behavior?
No.