fix(ci): test eicmkplugin with EICrecon_ROOT=install/ (#1164)
### Briefly, what does this PR introduce?
In CI, test the eicmkplugin with the local build, not system install. 

(Alternative is to modify eicmkplugin to write the install prefix into
the CMakeList as a HINT, but that requires adding too much to
eicmkplugin. It is also not a requirement so it would still fall back to
system installs.)

### What kind of change does this PR introduce?
- [x] Bug fix (issue #1163)
- [ ] New feature (issue #__)
- [ ] Documentation update
- [ ] Other: __

### Please check if this PR fulfills the following:
- [ ] Tests for the changes have been added
- [ ] Documentation has been added / updated
- [ ] Changes have been communicated to collaborators

### Does this PR introduce breaking changes? What changes might users
need to make to their code?
No.

### Does this PR change default behavior?
No.