fix(ci): don't upload unused build directory artifacts (#1123)
### Briefly, what does this PR introduce?
We spend several minutes just waiting for the build artifacts to get
uploaded, and then we don't use them... This only uploads the build
directory artifact that is used by the clang-tidy-iwyu job (which is,
incidentally, associated with the fasted compilation job, the smallest
build directory, and fastest upload). This will speed up when the build
matrix completes, and when the next set of jobs can start running.

### What kind of change does this PR introduce?
- [x] Bug fix (issue #__)
- [ ] New feature (issue #__)
- [ ] Documentation update
- [ ] Other: __

### Please check if this PR fulfills the following:
- [ ] Tests for the changes have been added
- [ ] Documentation has been added / updated
- [x] Changes have been communicated to collaborators @veprbl 

### Does this PR introduce breaking changes? What changes might users
need to make to their code?
No.

### Does this PR change default behavior?
No.