feat: migrate ParticlesWithPID_factory to JOmniFactory (#1178)
### Briefly, what does this PR introduce?
This migrates the ParticlesWithPID_factory to JOmniFactory. There are no
major issues here, but of note:
- the algorithm now returns tuple of unique_ptrs instead of struct,
- the algorithm now takes a single CherenkovParticleIDCollection as
third argument, instead of a vector of them (one for each Cherenkov
detector); this functionality can be restored if needed, or a separate
factory can merge the collections once we have pfRICH and DIRC
ParticleID collections,
- after defining the ParameterRefs, the parameters of this algorithm are
now parsed by JANA.

### What kind of change does this PR introduce?
- [ ] Bug fix (issue #__)
- [x] New feature (issue #__)
- [ ] Documentation update
- [ ] Other: __

### Please check if this PR fulfills the following:
- [ ] Tests for the changes have been added
- [ ] Documentation has been added / updated
- [ ] Changes have been communicated to collaborators

### Does this PR introduce breaking changes? What changes might users
need to make to their code?
No.

### Does this PR change default behavior?
No.