feat: use unique_ptr in MatchClusters; require all collections in reco.cc (#1175)
### Briefly, what does this PR introduce?
MatchClusters is trying to be clever in the factory, while at the same
time assuming hardcoded collection names. This changes the MatchClusters
factory to require MCParticles, ReconstructedChargedParticles, and
ReconstructedChargedParticleAssociations as first collections, then a
sequence of XClusters, XClusterAssociations (instead of building the
Associations listing itself). This makes the reco.cc more complete and
less dependent on MatchCluster_factory magic.

This also modifies MatchClusters the algorithm to return
tuple<unique_ptr<Collection> x 2>.

### What kind of change does this PR introduce?
- [ ] Bug fix (issue #__)
- [x] New feature (issue #__)
- [ ] Documentation update
- [ ] Other: __

### Please check if this PR fulfills the following:
- [ ] Tests for the changes have been added
- [ ] Documentation has been added / updated
- [ ] Changes have been communicated to collaborators

### Does this PR introduce breaking changes? What changes might users
need to make to their code?
No.

### Does this PR change default behavior?
No.

---------

Co-authored-by: pre-commit-ci[bot] <66853113+pre-commit-ci[bot]@users.noreply.github.com>
Co-authored-by: Dmitry Kalinkin <dmitry.kalinkin@gmail.com>