fix: convert TrackParamTruthInit to use DD4hep units (#1198)
### Briefly, what does this PR introduce?
Units are set (via command line or config files) assuming a certain unit
system. We have chosen DD4hep for this in all other cases (in part
because it comes with a parser). Having one algorithm use Acts units is
potentially confusing since distances may be passed in the wrong units.

This PR converts TrackParamTruthInit to store the config variables in
DD4hep units, and modifies the algorithm itself to interpret the config
in DD4hep units. No changes are expected to the output.

No values are specified in the factory or in the plugin, so this is the
only change that should be needed.

### What kind of change does this PR introduce?
- [x] Bug fix (issue #1166)
- [ ] New feature (issue #__)
- [ ] Documentation update
- [ ] Other: __

### Please check if this PR fulfills the following:
- [ ] Tests for the changes have been added
- [ ] Documentation has been added / updated
- [ ] Changes have been communicated to collaborators

### Does this PR introduce breaking changes? What changes might users
need to make to their code?
No.

### Does this PR change default behavior?
No.