Update HEXPLIT.cc
redefined the tolerance for finding the neighbors at specified positions so that it is relative to the side length.  This way it is loose enough that if there is a misalignment between layers in the detector it doesn't cause the HEXPLIT to not be able to determine the neighbors at each of these positions.