fix: explicitly set off-diagonal elements in the covariance matrices (#1244)
### Briefly, what does this PR introduce?
Acts uses square matrices for covarianc matrices that are **not**
required to be symmetric. This means that we should set **both**
off-diagonal elements explicitly (but we can just read one of them). The
EDM4eic data model does assume symmetry, so only upper diagonal is
stored.

This PR ensures that we explicitly write off-diagonal covariance matrix
elements everywhere, even if they are just written as zeros. This makes
it clearer that we are not implicitly assuming symmetry.

### What kind of change does this PR introduce?
- [x] Bug fix (issue: symmetry is assumed when not guaranteed)
- [ ] New feature (issue #__)
- [ ] Documentation update
- [ ] Other: __

### Please check if this PR fulfills the following:
- [ ] Tests for the changes have been added
- [ ] Documentation has been added / updated
- [ ] Changes have been communicated to collaborators

### Does this PR introduce breaking changes? What changes might users
need to make to their code?
No.

### Does this PR change default behavior?
No.